requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.66.0
pytest>=7.0.0
pandas>=2.0.0
matplotlib>=3.7.0